feat(Signup): Implemented the sign up and connect it to firebase
