+++ ThriftUni/app/(tabs)/signup.tsx
@@ -3,7 +3,20 @@ import { View } from '@/components/Themed';
 import React, { useState } from 'react';
 import { getAuth, createUserWithEmailAndPassword } from 'firebase/auth';
 import { getFirestore, doc, setDoc } from 'firebase/firestore';
-import { app } from '@/firebase'; // Asegúrate de que esté bien importado
+
+import { initializeApp } from "firebase/app";
+
+const firebaseConfig = {
+  apiKey: "AIzaSyD7EBLPezmrBujRy58eLzmAV1jeKTUrPoQ",
+  authDomain: "thriftuni-b345a.firebaseapp.com",
+  projectId: "thriftuni-b345a",
+  storageBucket: "thriftuni-b345a.firebasestorage.app",
+  messagingSenderId: "501062585933",
+  appId: "1:501062585933:web:7bb28a9b3f4f2f61a3d604",
+  measurementId: "G-MCWKZZQPTP"
+};
+
+const app = initializeApp(firebaseConfig);
 
 const auth = getAuth(app);
 const db = getFirestore(app);
